# Optimize GetControllerOf method and add a benchmark for it