@liggitt 2019-09-24
curious what led to optimizing here, was this showing up as a significant source of allocations?