@wojtek-t 2019-09-24
> curious what led to optimizing here, was this showing up as a significant source of allocations?

It was showing in both cpu and memory profile for controller-manager. It wasn't huge though (2-3% IIRC), but seemed like a really low hanging fruit.